+++ Models/SupplierPrefixes.cs
@@ -28,5 +28,6 @@ public static class SupplierPrefixes
 
     public static bool IsApiSupported(string supplier) =>
         supplier.Equals(Seco, StringComparison.OrdinalIgnoreCase) ||
-        supplier.Equals(Kennametal, StringComparison.OrdinalIgnoreCase);
+        supplier.Equals(Kennametal, StringComparison.OrdinalIgnoreCase) ||
+        supplier.Equals(Sandvik, StringComparison.OrdinalIgnoreCase);
 }